backend/pyproject.toml
[project]
name = "crossborder-erp-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.5",
    "django-environ>=0.13.0",
    "djangorestframework>=3.17.1",
    "drf-spectacular>=0.29.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-django>=4.12.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py"]
testpaths = ["apps", "tests"]
